Login Configuration
This patch adds default configuration for root and password logins and
makes it easy for users to overwrite these settings.

Logins as `root` are permitted by default using an SSH key, but not
using a password. While disabling this entirely is often a good idea,
settings that as default could easily lock out users by accident.

Using passwords for login is generally a bad idea and people should use
SSH keys anyway. Disabling this by default should not hurt.